src/agent/bridge/requirements.txt
# ZakOps Agent Bridge Dependencies
fastmcp>=2.14.0
httpx>=0.26.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
starlette>=0.35.0
